Sphinx documentation (#30)

+++ include/types.h
@@ -5,11 +5,18 @@
 #include <stddef.h>
 #include <stdbool.h>
 
+/** Register type */
 typedef volatile uintptr_t reg_t;
 
+/** Sleep type */
 typedef enum {
+	/** Only the core stopped while all peripherals continue to run */
 	SLEEP_NAP	= 1,
+	/** The core and peripherals stop to run including external clock while
+	 * still retaining the SRAM and register contents */
 	SLEEP_DEEP	= 2,
+	/** RTC and watchdog are the only ones continue to run. All the
+	 * contents of SRAM or the registers are not preserved */
 	SLEEP_BLACKOUT	= 3,
 } sleep_t;
 
@@ -50,7 +57,8 @@ typedef enum {
 #endif
 #endif
 
-static inline bool is_pow2(const unsigned int x)
+/** Check if integral power of two */
+static inline bool is_pow2(const uintptr_t x)
 {
 	return !(x & (x - 1));
 }
